% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linreg.r
\docType{class}
\name{linreg-class}
\alias{linreg}
\alias{linreg-class}
\title{Calculate a function for a multiple regression model}
\arguments{
\item{formula}{formula.}

\item{data}{data.frame.}
}
\value{
A linreg object.
}
\description{
Calculate a function for a multiple regression model
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula}

\item{\code{data}}{A data.frame}

\item{\code{bhat}}{A matrix}

\item{\code{resid}}{A matrix}

\item{\code{yhat}}{A matrix}

\item{\code{df}}{A integer}

\item{\code{rv1}}{A integer}

\item{\code{varcoeff}}{A matrix}

\item{\code{tstat}}{A matrix}

\item{\code{pv}}{A matrix}
}}
\examples{
lin1 <- linreg(formula = Petal.Length ~ Species, data = iris)
lin1$print()
lin1$summary()
}

