% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linreg.r
\docType{class}
\name{linreg-class}
\alias{linreg}
\alias{linreg-class}
\title{Calculate a function for a multiple regression model}
\value{
A linreg object.
}
\description{
Calculate a function for a multiple regression model
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}
}}
\examples{
lin1 <- linreg(formula = Petal.Length ~ Species, data = iris)
 lin1$print()
 lin1$plot()
}

