% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linreg.r
\docType{class}
\name{linreg-class}
\alias{linreg}
\alias{linreg-class}
\title{Calculate a function for a multiple regression model}
\arguments{
\item{formula}{formula.}

\item{data}{data.frame.}
}
\value{
A linreg object.
\usage{print(formula, data)
       plot(formula, data)
       resid(formula, data)
       pred(formula, data)
       coef(formula, data)
       summary(formula, data)
       }
}
\description{
Calculate a function for a multiple regression model
}
\examples{
{
lin1 <- linreg(formula = Petal.Length ~ Species, data = iris)
 lin1$print()
 lin1$plot()
 lin1$summary()
}
}

